requirements.txt
openai>=1.0.0
pytrends>=4.9.0
requests>=2.31.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pyyaml>=6.0
pillow>=10.0.0

# Web interface
flask>=3.0.0

# Image hosting
cloudinary>=1.36.0
